car filter fix on import year

+++ packages/plugin-tumentech-api/src/graphql/schema/tumentech.ts
@@ -217,8 +217,8 @@ const tumentechParams = `
   brand: String
   plateNumber: String
   vinNumber: String
-  vintageYear: Int
-  importYear: Int
+  vintageYear: Float
+  importYear: Float
   diagnosisDate: Date
   taxDate: Date
   drivingClassification: String
